task 1

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,3 +1,5 @@
+#include <stddef.h>
+
 /**
  * array_iterator - prints array
  * @array: inputs array
@@ -10,7 +12,7 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	int i;
+	size_t i;
 	
 	for (i = 0; i < size; i++)
 	{
